requirements.txt
scipy
numpy
Pillow
argparse
